LayoutTests/inspector/storage-panel-dom-storage-expected.txt

Test that storage panel is present and that it contains correct data for local and session DOM storages.

Populated local and session storage
Did show: Local storage
Did show: Session storage
Local storage content: 
KeyValuea0=value0, a6=value6, a8=value8, a4=value4, a2=value2, a1=value1, a9=value9, a3=value3, a7=value7, a5=value5, 
Session storage content: 
KeyValueb5=value15, b4=value14, b3=value13, b6=value16, b7=value17, b0=value10, b9=value19, b8=value18, b2=value12, b1=value11, 
DONE
